Add and handle exceptions
