+++ src/main/java/BotException.java
@@ -1,0 +1,7 @@
+import java.lang.RuntimeException;
+
+public class BotException extends RuntimeException {
+    public BotException(String message) {
+        super(message);
+    }
+}
