+++ src/main/java/Event.java
@@ -1,0 +1,17 @@
+public class Event extends Task {
+    protected String timing;
+
+    public Event(String desc, String timing) {
+        super(desc);
+        this.timing = timing;
+    }
+
+    @Override
+    public String toString() {
+        return "[E]" 
+                + super.toString()
+                + " (at: "
+                + this.timing
+                + ")";
+    }
+}
